Allow reference areas which extend past the canvas bounds (#1)
I’m working on a chart with several reference areas, some of which extend past the canvas bounds. Though I want to draw the part of the reference area that lies on the canvas, it doesn’t make sense to draw the part that spills out.

It's possible to grow the canvas to accommodate the reference area by specifying `alwaysShow`. Though after thinking about this for a while, it seems that _both_ these things should be possible: to show reference items which are partly off the canvas, and to extend the chart domain to accommodate reference items. This deprecates the `alwaysShow` prop, and adds an `ifOverflow` prop which supports four values:

- `ifOverflow="discard"` (default) – If the reference item falls partly outside the canvas, don't display it. This is the old default.
- `ifOverflow="extendDomain"` – If the reference item falls partly outside the canvas, extend the domain so it fits. This is the old `alwaysShow` behavior.
- `ifOverflow="clip"` – If the reference item falls partly outside the canvas, clip it to the canvas.
- `ifOverflow="keep"` – If the reference item falls partly outside the canvas, draw it anyway.