Use more liberal approach to dependency versions
This allows e.g. prop-types 15.7.0 to be used without issues. Without this change, many users will have libraries duplicated in their code because no versions are matching both e.g. ^15.7.0 and ~15.6.0 at the same time, or users will be forced not to update the libraries which happen to be also recharts dependencies.