Chore/lint staged (#3093)
Adding auto-linting to staged files using `husky` and `lint-staged`.

Removed `pre-commit` as I couldn't get it to run `lint-staged`.
`pre-push` was never working, add a hook that runs tests and linter on
push. - this is kind of annoying, open to feedback on it.

* run husky install on `prepare`
* run `lint-staged` on `pre-commit` (runs eslint and prettier)
* run `npm test` and `npm run lint` on `pre-push`

Also downgraded jest version to 27. Turns out 28 needs TypeScript 4.3+
[docs](https://jestjs.io/docs/28.x/upgrading-to-jest28)

Co-authored-by: Coltin Kifer <ckifer@amazon.com>