Storybook: Add Example section and remove redundant title from each story.  (#3329)
<!--- Provide a general summary of your changes in the Title above -->

## Description
In Storybook 7, the title of a story is optional. Without a title, the
folder structure is used to organise stories. Thus I removed the title
in every story.

Additionally I added the first example, to reproduce the structure of
API / Examples documentation as found on the website.

Finally, I remove a redundant folder structure, which is not needed,
because every component should have only a single file, with all stories
inside one file. Thus the component level folders are redundant.

With this structure we can now continue to merge all existing
documentation from (a) website, (b) demo and (c) the existing storybook
into this single place.
<!--- Describe your changes in detail -->

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context
This PR is another step towards a template of how to write and organise
stories. As the migration already started, I want to provide a structure
in which stories should be collected to guide and simplify the effort
required to migrate our documentation.

<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate):
<img width="371" alt="Screenshot 2023-02-05 at 17 35 58"
src="https://user-images.githubusercontent.com/27995606/216832011-97625979-2fde-44e4-a984-3d8e6b4efce7.png">

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.