fix: filter out functions in filterProps, remove type from valid SVGElementPropKeys (#3327)
<!--- Provide a general summary of your changes in the Title above -->

## Description

<!--- Describe your changes in detail -->
Recharts has a `filterProps` function which was intended to filter out
all properties that are not valid SVGProps or events that need passed
down to children.

Multiple times we have come across issues where Recharts defined props
such as `type` (in the past `points`) overlap with traditional SVGProps
(https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute)

In these cases props get spread onto SVG elements that aren't intended
or valid.

In this case the `type` prop [is a valid SVG
attribute](https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/type).
This never caused error in the past because the DOM will ignore
attributes that are strings or numbers, but when adding a line or area
`type` property that is custom `CurveFactory` from d3 such as
```
const stepAround = curveCardinal.tension(0.5);
```
the DOM will throw an error "Invalid value for prop type on <path> tag"
because functions are not valid on HTML elements in the DOM.

- remove `type` from valid `SVGElementAttributes` - we do not use any
elements that need it
- add an `isFunction` check within `filterProps` for attributes that are
supposed to be one of `SVGElementAttributes` - these should never be
functions
- detail comment within `filterProps`
- adds an example using a `CurveFactory` in AreaChart demo

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->
https://github.com/recharts/recharts/issues/3310

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->
Resolve above described error. Prevent more errors with `type` in the
future.

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

The true `type` SVG attribute can only be used with the following tags: 


![image](https://user-images.githubusercontent.com/25180830/216709142-dc93d508-de1b-43c6-a1be-0afeb3467a7a.png)

None of which we use in Recharts.

- In demo slightly modify an example to use a `CurveFactory` - ensure
the new curve factory type is displayed correctly and that the error is
not thrown.
- `type` is passed down explicitly by the chart so it is okay that it
gets filtered out in `filterProps`
- check functionality of other chart types and `CurveType` values such
as `monotone` and `linear`


## Screenshots (if appropriate):


![image](https://user-images.githubusercontent.com/25180830/216709982-a73e2436-2f2f-4332-a9dc-d6d4deb808fd.png)

^ No error thrown like in issue but curve still reflects the given
`CurveFactory`

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.

---------

Co-authored-by: Coltin Kifer <ckifer@amazon.com>
Co-authored-by: JM <proke03@naver.com>