fix: remove axis domain default prop and calculate it dynamically (#3328)
<!--- Provide a general summary of your changes in the Title above -->

## Description

<!--- Describe your changes in detail -->
Fixes #2593.

I removed the `domain` prop from the default props for all axis
components. As described in the issue comments, those default props do
not take the axis type into account and thereby use unsuitable domains
in some cases. This results in the bug.

Instead of using the component default props, the fallback domain is now
determined when the axis map is created. At this point the type of the
axis is known and is used to determine which fallback domain to use.

## Related Issue

<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please link to the issue here: -->

#2593

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->
Solves a bug.

## How Has This Been Tested?

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

I added a unit test for the BarChart component, as this specific case
was mentioned in the bug ticket.
I also tested charts with all the modified axis in the demo and
storybook. Both with and without specified domain.

## Screenshots (if appropriate):

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.