feat: add ability to set static domains to improve UX for toggling data (#3103)
Closes #3099 

This PR introduces the `includeHidden` prop to `Axis` elements. Setting
this to true ensures that all datapoints within a chart contribute to
its domain calculation, even when they are hidden. This allows
developers to improve the end-user experience for charts that allow
toggling datapoints, as it can improve their ability to make comparisons
on narrower sets of data. Documentation for this feature is added in
https://github.com/recharts/recharts.org/pull/223.

---------

Co-authored-by: Neef Rehman <Neef.Rehman@nationalgrid.com>